Changed folder structure to make use with Netlify easier.
