+++ src/components/Authentication/Authenticate.js
@@ -6,7 +6,7 @@ const Authenticate = (App) => class extends Component {
         super();
 
         this.state = {
-            loggedIn: true,
+            loggedIn: false,
         }
     }
